Unicode support!  Well, after a fashion:
Currently python_reader makes no attempt to work out the correct
encoding and UnixConsole defaults to 'utf-8'.  This is wrong, but it works
for me.

Reader.get_buffer returns a unicode string.  Ilpo & I decided a while back
that this is the wrong thing to do.

Fixes for both should follow shortly.